Use preallocated page for the console ring
get_console() is called before memory allocator is initialized. If
mapping it would require allocating new page (for page table on any
level), the kernel crashes.
It isn't strictly necessary with the current memory layout, but when
playing with linker script and memory layout (which the next patch do)
it is very easy to hit this issue, and suddenly debugging gets much
harder without console.